fix(worker): boot-shim heartbeats keep the startup watchdog honest under load
The startup watchdog is inactivity-based (re-arms on 'online' and every
message) — but between 'online' and the single final READY, a worker entry
evaluates its whole heavy import graph in silence. Under CPU contention
that silent evaluation exceeds the window while the worker is making
steady progress: the verify-gate flake, "Worker ready timeout after
10000ms of inactivity". No timeout values are changed by this fix.

The four worker entries (rsc/html × development/production) become thin
boot shims: static import only node:worker_threads (via a shared
bootWorker helper), post BOOTING immediately, heartbeat every second
(unref'd) while dynamically importing the implementation — module loading
yields to the event loop on file I/O, so beats flow while a
starved-but-alive worker loads. Public dist filenames are unchanged; the
implementations move to *.impl siblings and post READY exactly as before.

Required parent fix: createWorker attached its READY handler with
worker.once("message", ...), so the first heartbeat would consume it and
startup would never resolve — now on() with the existing self-removal.
A trailing beat in the READY handover window is ignored explicitly where
unknown message types warn.

Stress harness: new stub mode reproduces the boot-shim shape (busy chunks
with beats between, total load past the startup timeout, no silent gap
over the window) — fails with the once-listener, passes with on(). The
genuine-hang case (online, then silent) still fails as before.
